feat(spike.js): Client Class
